Break initial guess and coordinate mappings into seperate files
